+++ DeepQL/Environments/DiscreteEnv.cs
@@ -13,13 +13,13 @@ namespace DeepQL.Environments
         protected DiscreteEnv(int statesNum, int actionsNum)
             : base(new Discrete(actionsNum), new Discrete(statesNum))
         {
-            InitialStateDistribution = new double[statesNum];
+            InitialStateDistribution = new float[statesNum];
             TransitionsTable = new List<Transition>[statesNum, actionsNum];
             State = new Tensor(new Shape(1));
             LastAction = new Tensor(new Shape(1));
         }
 
-        public override bool Step(Tensor action, out Tensor observation, out double reward)
+        public override bool Step(Tensor action, out Tensor observation, out float reward)
         {
             var transitions = TransitionsTable[StateAsInt, (int)action[0]];
             int tInx = CategoricalSample(transitions.Select(x => x.Probability));
@@ -27,7 +27,7 @@ namespace DeepQL.Environments
             var t = transitions[tInx];
             LastActionAsInt = (int)action[0];
             StateAsInt = t.NextState;
-            observation = new Tensor(new double[] { t.NextState }, new Shape(1));
+            observation = new Tensor(new float[] { t.NextState }, new Shape(1));
             reward = t.Reward;
             return t.Done;
         }
@@ -39,30 +39,30 @@ namespace DeepQL.Environments
             return State.Clone();
         }
 
-        protected void SetInitialStateDistribution(int state, double weight)
+        protected void SetInitialStateDistribution(int state, float weight)
         {
             InitialStateDistribution[state] = weight;
         }
 
         protected void FinalizeInitialStateDistribution()
         {
-            double sum = InitialStateDistribution.Sum();
+            float sum = InitialStateDistribution.Sum();
 
             for (int i = 0; i < InitialStateDistribution.Length; ++i)
                 InitialStateDistribution[i] /= sum;
         }
 
-        protected void AddTransition(int state, int action, double probability, int nextState, double reward, bool done)
+        protected void AddTransition(int state, int action, float probability, int nextState, float reward, bool done)
         {
             if (TransitionsTable[state, action] == null) TransitionsTable[state, action] = new List<Transition>();
-            TransitionsTable[state, action].Add(new Transition() { Probability = 1.0, NextState = nextState, Reward = reward, Done = done });
+            TransitionsTable[state, action].Add(new Transition() { Probability = 1.0f, NextState = nextState, Reward = reward, Done = done });
         }
 
-        private int CategoricalSample(IEnumerable<double> probs)
+        private int CategoricalSample(IEnumerable<float> probs)
         {
             // assuming all probabilities sum up to 1
-            double p = Rng.NextDouble();
-            double probSum = 0;
+            float p = (float)Rng.NextDouble();
+            float probSum = 0;
             int probsNum = probs.Count();
 
             for (int i = 0; i < probsNum - 1; ++i)
@@ -77,16 +77,16 @@ namespace DeepQL.Environments
 
         private struct Transition
         {
-            public double Probability;
+            public float Probability;
             public int NextState;
-            public double Reward;
+            public float Reward;
             public bool Done;
         }
 
         public int StateAsInt { get { return (int)State[0]; } private set { State[0] = value; } }
         public int LastActionAsInt { get { return (int)LastAction[0]; } private set { LastAction[0] = value; } }
 
-        private readonly double[] InitialStateDistribution;
+        private readonly float[] InitialStateDistribution;
         private readonly List<Transition>[,] TransitionsTable;
     }
 }
